requirements.txt
opencv-python==4.8.1.78
moviepy==1.0.3
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
imageio==2.31.4
imageio-ffmpeg==0.4.9
scipy==1.11.3
matplotlib==3.7.2
ffmpeg-python==0.2.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1